update owl weight

+++ src/main/java/Owl.java
@@ -6,12 +6,12 @@ public class Owl {
 
     private  boolean hungry;
 
-    private double weight;
+    private int weight;
 
     private String name;
 
     //constructor
-    public Owl(int wingSpan, boolean hungry, double weight, String name) {
+    public Owl(int wingSpan, boolean hungry, int weight, String name) {
         this.wingSpan = wingSpan;
         this.hungry = hungry;
         this.weight = weight;
@@ -45,7 +45,7 @@ public class Owl {
         return weight;
     }
 
-    public void setWeight(double weight) {
+    public void setWeight(int weight) {
         if (weight >= 2000 && weight <= 4000) {
             this.weight = weight;
         }else {
